ref(onboarding): Split kotlin onboarding docs
