+++ static/app/gettingStartedDocs/kotlin/kotlin/onboarding.spec.tsx
@@ -2,7 +2,8 @@ import {renderWithOnboardingLayout} from 'sentry-test/onboarding/renderWithOnboa
 import {screen} from 'sentry-test/reactTestingLibrary';
 import {textWithMarkupMatcher} from 'sentry-test/utils';
 
-import docs, {PackageManager} from './kotlin';
+import {PackageManager} from './utils';
+import docs from '.';
 
 describe('GettingStartedWithKotlin', () => {
   it('renders gradle docs correctly', async () => {
